move DriveDesktopController to bapsf_motion.gui.configure.controllers
