+++ bapsf_motion/gui/configure/controllers.py
@@ -2,6 +2,7 @@
 
 import logging
 import numpy as np
+import re
 import warnings
 
 from abc import abstractmethod
@@ -15,6 +16,7 @@ from PySide6.QtWidgets import (
     QHBoxLayout,
     QGridLayout,
     QLayout,
+    QSizePolicy,
 )
 from typing import List
 
@@ -918,3 +920,237 @@ class DriveBaseController(QWidget):
             acw.close()
 
         event.accept()
+
+
+class DriveDesktopController(DriveBaseController):
+    def __init__(self, parent: QWidget | None = None):
+        super().__init__(axis_display_mode="interactive", parent=parent)
+
+    def _initialize_widgets(self):
+        # BUTTON WIDGETS
+        _btn = StyleButton("Move \n To", parent=self)
+        _btn.setMinimumHeight(100)
+        font = _btn.font()
+        font.setPointSize(20)
+        _btn.setFont(font)
+        self.move_to_btn = _btn
+
+        _btn = StyleButton("Home \n All", parent=self)
+        _btn.setMinimumHeight(100)
+        _btn.setFont(font)
+        _btn.setEnabled(False)
+        self.home_btn = _btn
+        self.home_btn.setVisible(False)
+
+        _btn = ZeroButton("Zero \n All", parent=self)
+        _btn.setMinimumHeight(100)
+        _btn.setFont(font)
+        self.zero_all_btn = _btn
+
+        _btn = StyleButton("Holding\nCurrent", parent=self)
+        _btn.setFixedHeight(44)
+        font = _btn.font()
+        font.setPointSize(10)
+        _btn.setFont(font)
+        _btn.update_style_sheet(
+            styles={
+                "background-color": re.sub(
+                    " +",
+                    " ",
+                    """qlineargradient(
+                        x1:0,
+                        y1:0, 
+                        x2:1, 
+                        y2:0,
+                        stop: 0 rgb(52, 161, 219),
+                        stop: 0.1 rgb(52, 161, 219),
+                        stop: 0.4 rgb(163, 163, 163),
+                        stop: 1 rgb(163, 163, 163)
+                    )""".replace("\n", ""),
+                ),
+            },
+            action="base",
+        )
+        _btn.update_style_sheet(
+            styles={
+                "background-color": re.sub(
+                    " +",
+                    " ",
+                    """qlineargradient(
+                        x1:0,
+                        y1:0, 
+                        x2:1, 
+                        y2:0,
+                        stop: 0 rgb(163, 163, 163),
+                        stop: 0.6 rgb(163, 163, 163),
+                        stop: 0.9 rgb(250, 66, 45)
+                        stop: 1 rgb(250, 66, 45)
+                    )""".replace("\n", ""),
+                ),
+            },
+            action="checked",
+        )
+        _btn.setCheckable(True)
+        _btn.setChecked(False)
+        self.hold_current_btn = _btn
+
+    def _connect_signals(self):
+        super()._connect_signals()
+
+        self.zero_all_btn.clicked.connect(self.zeroDrive.emit)
+        self.move_to_btn.clicked.connect(self._move_to)
+        self.hold_current_btn.clicked.connect(self._toggle_holding_current)
+
+    def _define_layout(self) -> QLayout:
+        _on = QLabel("O\nN", parent=self)
+        font = _on.font()
+        font.setBold(True)
+        _on.setFont(font)
+        _on.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        _on.setFixedWidth(10)
+        _on.setSizePolicy(QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Minimum)
+
+        _off = QLabel("O\nF\nF", parent=self)
+        _off.setFont(font)
+        _off.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        _off.setFixedWidth(10)
+        _off.setSizePolicy(QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Minimum)
+
+        holding_current_layout = QHBoxLayout()
+        holding_current_layout.setContentsMargins(0, 0, 0, 0)
+        holding_current_layout.addWidget(
+            _on,
+            alignment=Qt.AlignmentFlag.AlignVCenter,
+        )
+        holding_current_layout.addWidget(self.hold_current_btn)
+        holding_current_layout.addWidget(
+            _off,
+            alignment=Qt.AlignmentFlag.AlignVCenter,
+        )
+        _on.setVisible(False)
+        _off.setVisible(False)
+        self.hold_current_btn.setVisible(False)
+
+        # Sub-Layout #1
+        sub_layout = QVBoxLayout()
+        sub_layout.setContentsMargins(0, 0, 0, 0)
+        sub_layout.addWidget(self.move_to_btn)
+        sub_layout.addStretch()
+        # sub_layout.addWidget(self.home_btn)
+        sub_layout.addLayout(holding_current_layout)
+        sub_layout.addStretch()
+        sub_layout.addWidget(self.zero_all_btn)
+        sub_widget = QWidget(parent=self)
+        sub_widget.setLayout(sub_layout)
+        sub_widget.setFixedWidth(140)
+
+        # Sub-Layout #2
+        _text = QLabel("Position", parent=self)
+        font = _text.font()
+        font.setPointSize(14)
+        _text.setFont(font)
+        _pos_label = _text
+
+        _text = QLabel("Target", parent=self)
+        font = _text.font()
+        font.setPointSize(14)
+        _text.setFont(font)
+        _target_label = _text
+
+        _text = QLabel("Jog Δ", parent=self)
+        font = _text.font()
+        font.setPointSize(14)
+        _text.setFont(font)
+        _jog_delta_label = _text
+
+        sub_layout2 = QVBoxLayout()
+        sub_layout2.setSpacing(8)
+        sub_layout2.addSpacing(54)
+        sub_layout2.addWidget(
+            _pos_label,
+            alignment=Qt.AlignmentFlag.AlignVCenter | Qt.AlignmentFlag.AlignRight,
+        )
+        sub_layout2.addSpacing(42)
+        sub_layout2.addWidget(
+            _target_label,
+            alignment=Qt.AlignmentFlag.AlignVCenter | Qt.AlignmentFlag.AlignRight,
+        )
+        sub_layout2.addSpacing(86)
+        sub_layout2.addWidget(
+            _jog_delta_label,
+            alignment=Qt.AlignmentFlag.AlignVCenter | Qt.AlignmentFlag.AlignRight,
+        )
+        sub_layout2.addStretch(1)
+
+        layout = QHBoxLayout()
+        layout.addWidget(sub_widget)
+        layout.addLayout(sub_layout2)
+        for acw in self._axis_control_widgets:
+            layout.addWidget(acw)
+            layout.addSpacing(2)
+        layout.addStretch()
+
+        return layout
+
+    @Slot()
+    def _move_to(self):
+        target_pos = [
+            acw.target_position
+            for acw in self._axis_control_widgets
+            if not acw.isHidden()
+        ]
+
+        if self.mg.drive.is_moving:
+            self.logger.info(
+                "Probe drive is currently moving.  Did NOT perform move "
+                f"to {target_pos}."
+            )
+            target_pos = []
+
+        if any(p is None for p in target_pos):
+            self.logger.warning(
+                f"Requested target position ({target_pos}) is not valid,"
+                f" NOT performing move to."
+            )
+            return
+
+        self.moveTo.emit(target_pos)
+
+    @Slot()
+    def _toggle_holding_current(self):
+        hold_current = not self.hold_current_btn.isChecked()
+        if hold_current:
+            self.mg.drive.send_command("reset_currents")
+        else:
+            idle_currents = [0] * self.mg.drive.naxes
+            self.mg.drive.send_command("set_idle_current", *idle_currents)
+
+    def set_target_position(self, target_position: List[float]):
+        npos = len(target_position)
+        naxes = self.mg.drive.naxes
+
+        if npos != naxes:
+            self.logger.warning(
+                f"Received target position {target_position} does NOT "
+                f"have the same dimensionality as the drive "
+                f"({naxes})."
+            )
+            return
+
+        for ii, pos in enumerate(target_position):
+            acw = self._axis_control_widgets[ii]
+            acw.update_target_position_display(pos)
+
+    def disable_motion_buttons(self):
+        self.move_to_btn.setEnabled(False)
+        self.zero_all_btn.setEnabled(False)
+        self.hold_current_btn.setEnabled(False)
+
+        super().disable_motion_buttons()
+
+    def enable_motion_buttons(self):
+        self.move_to_btn.setEnabled(True)
+        self.zero_all_btn.setEnabled(True)
+        self.hold_current_btn.setEnabled(True)
+
+        super().enable_motion_buttons()
